WIP: Accommodate the FreeRTOS port primitives defined by HALCoGen.
HALCoGen emits a FreeRTOS 9.0.0 codebase and enables some different options than
upstream FreeRTOS 10.4.6's RM4 port does. (Most notably, HALCoGen's version
enables the MPU.) This is an attempt to add the necessary port implementations
to support the MPU and some extra primitives that upstream FreeRTOS expects.

This is not yet completely working and we're putting it on hold for now.